requirements.txt
# Core LLM & Orchestration
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0

# Document Indexing & RAG
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0

# Vector Database
chromadb>=0.5.0

# File Parsers
pdfplumber>=0.11.0
python-docx>=1.1.0
python-pptx>=0.6.23

# Data Validation
pydantic>=2.7.0
pydantic-settings>=2.3.0

# UI
streamlit>=1.36.0
streamlit-extras>=0.4.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.3.0
loguru>=0.7.0
pandas>=2.2.0
plotly>=5.22.0